Configuring files

+++ eslint.config.mjs
@@ -19,7 +19,7 @@ export default [
     rules: {
       // ✅ Style / Safety
       semi: ['error', 'always'],
-      quotes: ['error', 'single', { avoidEscape: true }],
+      quotes: ['error', 'double', { avoidEscape: true }],
       'no-console': ['warn', { allow: ['warn', 'error'] }],
       'no-unused-vars': 'warn',
       'prefer-const': 'error',
